+++ desktop/appcontainer/ToastSource.ts
@@ -4,13 +4,15 @@
  *
  * Copyright © 2023 Extremely Heavy Industries Inc.
  */
+import {OverlayToasterProps} from '@blueprintjs/core/src/components/toast/toaster';
 import {ToastModel} from '@xh/hoist/appcontainer/ToastModel';
 import {ToastSourceModel} from '@xh/hoist/appcontainer/ToastSourceModel';
 import {div} from '@xh/hoist/cmp/layout';
-import {hoistCmp, HoistModel, useLocalModel, uses, lookup} from '@xh/hoist/core';
+import {elementFactory, hoistCmp, HoistModel, lookup, useLocalModel, uses} from '@xh/hoist/core';
 import {OverlayToaster, ToasterPosition} from '@xh/hoist/kit/blueprint';
 import classNames from 'classnames';
 import {isElement, map} from 'lodash';
+import {createRoot} from 'react-dom/client';
 import {wait} from '../../promise';
 import './Toast.scss';
 
@@ -41,20 +43,20 @@ class ToastSourceLocalModel extends HoistModel {
         const {sourceModel} = this;
         this.addReaction({
             track: () => [sourceModel.toastModels, map(sourceModel.toastModels, 'isOpen')] as const,
-            run: ([models]) => this.displayPendingToasts(models)
+            run: ([models]) => this.displayPendingToastsAsync(models)
         });
     }
 
-    displayPendingToasts(models: ToastModel[]) {
-        models.forEach((model: ToastModel & {bpId}) => {
+    async displayPendingToastsAsync(models: ToastModel[]) {
+        for (const model: ToastModel & {bpId} of models) {
             let {bpId, isOpen, icon, intent, actionButtonProps, position, containerRef, ...rest} =
                 model;
 
             // 1) If toast is visible and sent to bp, or already obsolete -- nothing to do
-            if (!!bpId === isOpen) return;
+            if (!!bpId === isOpen) continue;
 
             // 2) ...otherwise this toast needs to be shown or hidden with bp api
-            let toaster = this.getToaster(position as ToasterPosition, containerRef);
+            let toaster = await this.getToasterAsync(position as ToasterPosition, containerRef);
             if (!bpId) {
                 model.bpId = toaster.show({
                     className: classNames('xh-toast', `xh-toast--${intent}`),
@@ -67,7 +69,7 @@ class ToastSourceLocalModel extends HoistModel {
             } else {
                 toaster.dismiss(bpId);
             }
-        });
+        }
     }
 
     /**
@@ -80,7 +82,7 @@ class ToastSourceLocalModel extends HoistModel {
      * @param position - position on screen where toast should appear.
      * @param containerRef - DOM Element used to position (contain) the toast.
      */
-    getToaster(position: ToasterPosition, containerRef: HTMLElement) {
+    async getToasterAsync(position: ToasterPosition, containerRef: HTMLElement) {
         if (containerRef && !isElement(containerRef)) {
             console.warn(
                 'containerRef argument for Toast must be a DOM element. Argument will be ignored.'
@@ -94,8 +96,37 @@ class ToastSourceLocalModel extends HoistModel {
         // We want to just memoize this by two args (one object)?  Is there a library for this?
         const toasters = toasterMap.get(container) || {};
         if (!toasters[position])
-            toasters[position] = OverlayToaster.create({position, className}, container);
+            toasters[position] = await this.createToaster({position, className}, container);
         toasterMap.set(container, toasters);
         return toasters[position];
     }
+
+    /** Work around to avoid calling OverlayToaster.create(). It uses ReactDOM.render
+     * that gives a warning because it is deprecated in React 18.
+
+     * The use of ReactDOM.render set to be removed from OverlayToaster.create() in Blueprint v6.0
+     * https://github.com/palantir/blueprint/issues/5212#issuecomment-1294958195
+     * */
+    private createToaster(props?: OverlayToasterProps, container = document.body) {
+        const containerElement = document.createElement('div');
+        container.appendChild(containerElement);
+        const root = createRoot(containerElement);
+        return new Promise<OverlayToaster>((resolve, reject) => {
+            root.render(
+                overlayToaster({
+                    ...props,
+                    usePortal: false,
+                    ref: instance => {
+                        if (instance) {
+                            resolve(instance);
+                        } else {
+                            reject(new Error('[Blueprint] Unable to create toaster.'));
+                        }
+                    }
+                })
+            );
+        });
+    }
 }
+
+const overlayToaster = elementFactory(OverlayToaster);