+++ CHANGELOG.md
@@ -32,6 +32,10 @@
 * Fixed bug where `manuallySized` was not being set properly on column state
 * Grid header bottom border still rendering when hiding grid headers fix.
 
+### ⚙️ Technical
+
+* Prevent Blueprint Toaster from triggering ReactDOM.render console warnings
+
 ### 📚 Libraries
 
 * mobx `6.8 -> 6.9`